refactor: clean up code to production level
