+++ app/src/main/java/ru/popkov/marvelapp/MainWindow.kt
@@ -2,10 +2,7 @@ package ru.popkov.marvelapp
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
@@ -22,16 +19,13 @@ fun MainWindow(
     val items = navProviders
         .mapNotNull { it.item }
         .sortedBy { it.index }
-    val snackbarHostState = remember { SnackbarHostState() }
 
     NavigationLaunchedEffect(
         navigator = navigator,
         navController = navController,
     )
 
-    Scaffold(
-        snackbarHost = { SnackbarHost(snackbarHostState) },
-    ) { innerPadding ->
+    Scaffold { innerPadding ->
         NavHost(
             navController = navController,
             startDestination = (items.firstOrNull { it.isStart } ?: items.first()).route,
@@ -39,7 +33,7 @@ fun MainWindow(
                 .padding(innerPadding),
         ) {
             navProviders.forEach { subGraph ->
-                subGraph.graph(this, snackbarHostState)
+                subGraph.graph(this)
             }
         }
     }
