+++ core/feature/nav/src/main/java/ru/popkov/android/core/feature/nav/NavProvider.kt
@@ -1,6 +1,5 @@
 package ru.popkov.android.core.feature.nav
 
-import androidx.compose.material3.SnackbarHostState
 import androidx.navigation.NavGraphBuilder
 
 interface NavProvider {
@@ -15,6 +14,5 @@ interface NavProvider {
 
     fun graph(
         builder: NavGraphBuilder,
-        snackbarHostState: SnackbarHostState,
     )
 }
