+++ theme/src/main/res/values/themes.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<resources>
+<resources xmlns:tools="http://schemas.android.com/tools">
 
-    <style name="Theme.MarvelApp" parent="android:Theme.Material.Light.NoActionBar" />
+    <style name="Theme.MarvelApp" parent="android:Theme.Material.Light.NoActionBar">
+        <item name="android:forceDarkAllowed" tools:targetApi="q">false</item> // For MIUI
+    </style>
 </resources>
